Split Preprocessor CLI and Command Modules
Move the existing clap command surface into cli.rs, root resolution into
context.rs, and build layout constants into layout.rs.

Move changed-file execution and the shared is_proposal_path classifier
into changed.rs while preserving existing runtime behavior.

Keep main.rs focused on top-level orchestration and add initial
module/test ownership notes for later preprocessor slices.
