+++ src/context.rs
@@ -1,0 +1,20 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+use std::path::PathBuf;
+
+use snafu::{ResultExt, Whatever};
+
+use crate::{cli::Args, find_root};
+
+pub(crate) fn root(args: &Args) -> Result<PathBuf, Whatever> {
+    let dir = match &args.root {
+        None => find_root::find_root().whatever_context("cannot find repository root")?,
+        Some(p) => p.to_path_buf(),
+    };
+    find_root::is_root(&dir).whatever_context("invalid root directory")?;
+    Ok(dir)
+}
