+++ src/layout.rs
@@ -1,0 +1,10 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+pub(crate) const CONTENT_DIR: &str = "content";
+pub(crate) const BUILD_DIR: &str = "build";
+pub(crate) const REPO_DIR: &str = "repo";
+pub(crate) const OUTPUT_DIR: &str = "output";
